+++ src/main/scala/com/guizmaii/distances/implementations/cache/RedisGeoCache.scala
@@ -1,17 +1,23 @@
 package com.guizmaii.distances.implementations.cache
 
+import org.apache.commons.pool2.impl.GenericObjectPoolConfig
+import redis.clients.jedis.JedisPool
+
 import scala.concurrent.duration.Duration
 import scala.reflect.ClassTag
 
-final class RedisGeoCache[E <: Serializable: ClassTag](host: String, port: Int, override val expiration: Duration) extends GeoCache[E] {
+final class RedisGeoCache[E <: Serializable: ClassTag](jedisPool: JedisPool, override val expiration: Duration) extends GeoCache[E] {
   import scalacache._
   import scalacache.redis._
   import scalacache.serialization.binary._
 
-  override implicit val innerCache: Cache[E] = RedisCache[E](host, port)
+  override implicit val innerCache: Cache[E] = RedisCache[E](jedisPool)
 }
 
 object RedisGeoCache {
   def apply[E <: Serializable: ClassTag](host: String, port: Int, expiration: Duration): RedisGeoCache[E] =
-    new RedisGeoCache(host, port, expiration)
+    new RedisGeoCache(new JedisPool(host, port), expiration)
+
+  def apply[E <: Serializable: ClassTag](host: String, port: Int, password: String, expiration: Duration): RedisGeoCache[E] =
+    new RedisGeoCache(new JedisPool(new GenericObjectPoolConfig(), host, port, 1000, password), expiration)
 }